// src/app/interfaces/customer.ts
export interface Customer {
  idcustomer: string;
  taxnumber: string;
  lastname: string;
  firstname: string;
  username: string;
  password: string;
  secquestion: string;
  secanswer: string;
  registered: string;
  registrationtoken: string;
  subscribed: string;
  gender: string;
  dateofbirth: any;
  placeofbirth: string;
  address: string;
  zipcode: string;
  city: string;
  country: string;
  mobile: string;
  email: string;
  active: string;
  changedby: string;
  changedat: string;
}